correcciones y pruebas del service hechas

+++ EduProveedores/src/main/java/com/edutech/EduProveedores/repository/ProveedorRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.edutech.EduProveedores.model.Proveedor;
 
-public interface ProveedorRepository extends JpaRepository <Proveedor, Integer> {
+public interface ProveedorRepository extends JpaRepository <Proveedor, Long> {
     
 
 }
